msm: vidc: Fix broken debugfs creation error checks and error paths
edit: modify for older version in this source

The debugfs helper functions do not always return NULL when they fail;
instead, they can return an error number casted as a pointer, so that their
users have the option to determine the exact cause of failure.

Use the IS_ERR_OR_NULL() helper when checking for debugfs errors to fix the
error checks.

Signed-off-by: Sultanxda <sultanxda@gmail.com>